Major refactor to use hh-util and normalize build for parenting projects:
* nfg->hh or hangouthere nomenclature changes
* Created localized compose scripts:
	* compose/dev-linked.sh: Similar to the hh-dev-linked will establish a
	  triggerWatch, as well as initiate an hh-dev
	* compose/publish.sh: Execs docker cli to call hh-publish for us
* Removed previous .scripts in favor of those bundled with hh-util
* Rewrite Dockerfile to be more simplistic and default to a prod build
  env
	* compose files will change appropriate ENV vars to development
	  if needed
* docker-compose.dev_linked.yml:
	* Now handles vol ownership changeroo
	* Command updated to run gh-compose-dev-linked similarly to
	  hh-util project
* Added compose test file for consistency and global functionality with
  hh-compose-* scripts
* docker-compose.update-deps.yml updated to match hh-util stanzas
* package.json:
	* Heavy updates removing npm scripts in favor of hh-util bin
	  scripts
	* Tiny updates for localized "gh" bin scripts
* publishConfig updated to not use `git tag` while versioning
* Update some of the typescript to use modern ES2015 modules instead of
  TS namespaces. Was causing an eslint and tsc conflicting error.
