+++ .scripts/compose/publish.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+# Publish the lib to the targeted registry
+
+docker exec -it golden-hammer-shared npx hh-publish gh-shared
