--- .scripts/run-dev.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-# Run project in standard development mode
-
-docker compose \
-  -f ./docker-compose.yml \
-  -f ./docker-compose.dev.yml \
-  up
