--- .scripts/run-prod.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-
-# Run a Production Build against source
-
-docker compose \
-  -f ./docker-compose.yml \
-  -f ./docker-compose.prod.yml \
-  up
