--- .scripts/run-update-deps.sh
@@ -1,5 +1,0 @@
-#!/bin/sh
-
-docker compose \
-  -f ./docker-compose.update-deps.yml \
-  up
