+++ Dockerfile
@@ -1,151 +1,27 @@
-ARG PATH_APP=/app
+FROM node:18-alpine3.18
 
-##########################################################################################################################################
-# Stage: Base
-# - Creates necessary paths
+ENV NODE_ENV=production
 
-FROM node:18.8.0-alpine3.15 AS base
+RUN apk add --no-cache \
+  g++ \
+  gcc \
+  git \
+  libgcc \
+  libstdc++ \
+  linux-headers \
+  make \
+  python3 \
+  sqlite && \
+  mkdir /app && chown -R node.node /app
 
-  # Import into Stage
-  ARG PATH_APP
+WORKDIR /app
 
-  USER root
+USER node
 
-  # Make necessary Directories, and chown them as node for dev build path
-  RUN mkdir -p $PATH_APP $RUN_PATH && \
-    chown node:node $PATH_APP $RUN_PATH
+COPY --chown=node:node package*.json ./
 
-##########################################################################################################################################
-# Stage: Base Build OS
-# - Includes all base packages from OS needed to build
+RUN npm install --production
 
-FROM base AS build-base
+COPY --chown=node:node . .
 
-  ENV PYTHON=/usr/bin/python2
-
-  # Node Compilation stuff
-  RUN apk add --no-cache \
-    g++ \
-    gcc \
-    git \
-    libgcc \
-    libstdc++ \
-    linux-headers \
-    make \
-    python3 \
-    sqlite
-
-##########################################################################################################################################
-# Stage: Development Environment
-# - Installs npm dependencies in development mode
-
-FROM build-base AS development
-
-  # Import into Stage
-  ARG PATH_APP
-
-  ENV NODE_ENV=development
-
-  WORKDIR ${PATH_APP}
-
-  USER node
-
-  COPY --chown=node:node . ./
-
-  RUN npm i --verbose
-
-  CMD ["npx", "nfg-compose"]
-
-# ! FIXME: This was the original, but we're modifying while we figure out the build process
-
-# FROM build-base AS developmentORIG
-
-#   # Import into Stage
-#   ARG PATH_APP
-
-#   ENV NODE_ENV=development
-
-#   WORKDIR ${PATH_APP}
-
-#   # Fixes issue with image running as root, but npm script running as `node`
-#   # See: comment block at bottom of this Stage
-#   RUN npm config set home /home/node/
-#   RUN npm config set cache /home/node/.npm
-
-#   COPY ./package*.json ./
-
-#   # legacy-peer-deps required because libs are dumb, not updating peers
-#   RUN npm i --legacy-peer-deps
-
-#   # Copy the source to allow development
-#   COPY . ./
-
-#   # At this point, source is built and owned by root inside the image,
-#   # but mostly as a transition to the `production-build` Stage.
-#   #
-#   # It's important to point out the assumption that while developing,
-#   # the volume mounted at /build will be owned by your host UID 1000,
-#   # which causes npm to run as the user `node` inside the
-#   # image (also UID 1000).
-#   # See: https://docs.npmjs.com/cli/v8/using-npm/scripts#user
-#   #
-#   # Additionally, files must be owned by root for Bitbucket file transfers
-#   # between docker build Stages, due to the way they manage userns
-#   # See: https://github.com/moby/moby/issues/34645
-
-#   CMD ["npm", "run", "docker-compose"]
-
-##########################################################################################################################################
-# Stage: Production Build
-# - Rebuilds the project in Production Mode
-
-FROM development AS production-build
-
-      # Import into Stage
-      ARG PATH_APP
-
-      ENV NODE_ENV=production
-
-      WORKDIR ${PATH_APP}
-
-      COPY --from=development ${PATH_APP} .
-
-      # Re-build in Prod Mode
-      RUN npm run clean
-      RUN npm run build
-
-      # Re-build modules in Prod Mode
-      RUN rm -rf node_modules
-      RUN npm ci --legacy-peer-deps
-
-##########################################################################################################################################
-# Stage: Production Environment
-# - Copies build to clean node env
-
-FROM base AS production
-
-      # Import into Stage
-      ARG PATH_APP
-
-      ENV NODE_ENV=production
-
-      WORKDIR ${PATH_APP}
-
-      # Copy package manifest and build from `production-build`
-      COPY --chown=node:node ./package*.json ./
-
-      COPY --from=production-build --chown=node:node ${PATH_APP}/dist/ ./
-      COPY --from=production-build --chown=node:node ${PATH_APP}/node_modules ./node_modules
-
-      RUN chown -R node:node /home/node
-
-      # Quick hack to disable update notifications from `npm` on runs
-      USER node
-      RUN npm config set update-notifier false
-      # Ahhh, back to root
-      USER root
-
-      RUN npm config set update-notifier false
-
-      ENTRYPOINT ["/bin/sh"]
-      CMD ["npm", "start"]
+CMD ["npm", "start"]
