+++ docker-compose.dev.yml
@@ -2,8 +2,9 @@ version: '3.7'
 
 services:
   golden-hammer-shared:
-    build:
-      target: development
-    command: ['npm', 'run', 'dev']
+    container_name: golden-hammer-shared
+    command: npx hh-dev
     environment:
+      - NODE_ENV=development
+      - VERBOSE=1
       - NODE_OPTIONS=--enable-source-maps --trace-warnings
