+++ docker-compose.test.yml
@@ -2,5 +2,4 @@ version: '3.7'
 
 services:
   golden-hammer-shared:
-    environment:
-      - TZ=America/Chicago
+    command: ''
