+++ docker-compose.yml
@@ -2,15 +2,13 @@ version: '3.7'
 
 services:
   golden-hammer-shared:
-    image: nfg/golden-hammer-shared:latest
+    image: hangouthere/golden-hammer-shared:latest
     build:
       context: .
       dockerfile: ./Dockerfile
-      target: development
     stdin_open: true
     tty: true
+    user: '1000'
+    working_dir: '/app'
     volumes:
       - .:/app
-      # - /app/node_modules
-    # environment:
-    #   - TERM=xterm-256color
